ws: Remove most of the polkit and reauthorization code
This is our previous reauthorization logic which is being
replaced with the stock polkit agent.

Closes #6081
