+++ containers/kubernetes/install.sh
@@ -23,8 +23,6 @@ rm -rf /usr/share/cockpit/realmd/ /usr/share/cockpit/systemd/ /usr/share/cockpit
 
 # Remove unwanted cockpit-bridge binaries
 rm -rf /usr/bin/cockpit-bridge
-rm -rf /usr/lib64/security/pam_reauthorize.so
-rm -rf /usr/libexec/cockpit-polkit
 rm -rf /usr/libexec/cockpit-askpass
 
 rm -rf /container/scripts
