--- src/reauthorize/HACKING
@@ -1,14 +1,0 @@
-
-Working on pam_reauthorize.so
-------------------------------
-
-For now this is part of Cockpit, but may eventually be split out. So only
-basic glibc and linux dependencies are allowed:
-
- * glibc
- * keyutils
- * pam
-
-To run some of the test you'll need to:
-
- $ sudo make enable-pam-tests
