--- src/reauthorize/Makefile-reauthorize.am
@@ -1,64 +1,0 @@
-
-noinst_LIBRARIES += libreauthorize.a
-
-libreauthorize_a_SOURCES = \
-	src/reauthorize/reauthorize.c \
-	src/reauthorize/reauthorize.h \
-	$(NULL)
-libreauthorize_a_LIBS = \
-	libreauthorize.a \
-	$(REAUTHORIZE_LIBS)
-
-pam_reauthorize_FILES = src/reauthorize/pam_reauthorize.c src/reauthorize/reauthorize.c
-pam_reauthorize.so: $(pam_reauthorize_FILES)
-	$(AM_V_CCLD) $(CC) -fPIC -shared $(CFLAGS) -I$(builddir) -o $@ $^ $(REAUTHORIZE_LIBS) $(LDFLAGS)
-
-pam_MODULES = pam_reauthorize.so
-
-all-local:: $(pam_MODULES)
-
-EXTRA_DIST += \
-	$(pam_reauthorize_FILES) \
-	$(NULL)
-
-CLEANFILES += $(pam_MODULES)
-
-install-exec-local::
-	$(MKDIR_P) $(DESTDIR)$(pamdir)
-	$(INSTALL) pam_reauthorize.so $(DESTDIR)$(pamdir)
-uninstall-local::
-	$(RM) -f $(DESTDIR)$(pamdir)/pam_reauthorize.so
-
-# -----------------------------------------------------------------------------
-# Tests
-
-noinst_PROGRAMS += \
-	mock-reauthorize \
-	$(NULL)
-
-mock_reauthorize_SOURCES = src/reauthorize/mock-reauthorize.c
-mock_reauthorize_CFLAGS = \
-	-DSYSCONFDIR=\"$(sysconfdir)\" \
-	-DBUILDDIR=\"$(abs_builddir)\"
-	$(NULL)
-mock_reauthorize_LDADD = $(libreauthorize_a_LIBS)
-
-REAUTHORIZE_CHECKS = \
-	test-reauthorize \
-	$(NULL)
-
-test_reauthorize_CFLAGS = $(libreauthorize_a_CFLAGS)
-test_reauthorize_SOURCES = src/reauthorize/test-reauthorize.c
-test_reauthorize_LDADD = libretest.a $(libreauthorize_a_LIBS)
-
-noinst_PROGRAMS += $(REAUTHORIZE_CHECKS)
-TESTS += $(REAUTHORIZE_CHECKS)
-
-EXTRA_DIST += $(pam_CONFIGS)
-
-enable-root-tests:
-	chown 0:0 $(builddir)/mock-reauthorize
-	chmod u+s $(builddir)/mock-reauthorize
-
-disable-pam-tests:
-	chmod u-s $(builddir)/mock-reauthorize
