+++ test/images/scripts/debian-8.setup
@@ -16,8 +16,6 @@ COCKPIT_DEPS="\
 cryptsetup \
 libjson-glib-1.0-0 \
 libpcp3 \
-libpolkit-agent-1-0 \
-libpolkit-gobject-1-0 \
 libpwquality-tools \
 libssh-4 \
 libvirt-daemon-system \
@@ -48,10 +46,8 @@ libssh-dev \
 zlib1g-dev \
 libkrb5-dev \
 libxslt1-dev \
-libkeyutils-dev \
 libglib2.0-dev \
 libsystemd-dev \
-libpolkit-agent-1-dev \
 libpcp3-dev \
 libjson-glib-dev \
 libpam0g-dev \
