+++ test/images/scripts/debian-testing.setup
@@ -15,8 +15,7 @@ COCKPIT_DEPS="\
 cryptsetup \
 docker.io \
 libjson-glib-1.0-0 \
-libpolkit-agent-1-0 \
-libpolkit-gobject-1-0 \
+libpcp3 \
 libpwquality-tools \
 libssh-4 \
 libteam-utils \
@@ -47,10 +46,9 @@ libssh-dev \
 zlib1g-dev \
 libkrb5-dev \
 libxslt1-dev \
-libkeyutils-dev \
 libglib2.0-dev \
 libsystemd-dev \
-libpolkit-agent-1-dev \
+libpcp3-dev \
 libjson-glib-dev \
 libpam0g-dev \
 systemd \
