+++ tools/cockpit.debian.pam
@@ -1,7 +1,6 @@
 #%PAM-1.0
 auth	   required	pam_sepermit.so
 auth       substack     common-auth
-auth       optional     pam_reauthorize.so prepare
 auth       optional     pam_ssh_add.so
 account    required     pam_nologin.so
 account    required     pam_shells.so
@@ -13,7 +12,6 @@ session    required     pam_loginuid.so
 # pam_selinux.so open should only be followed by sessions to be executed in the user context
 session    required     pam_selinux.so open env_params
 session    optional     pam_keyinit.so force revoke
-session    optional     pam_reauthorize.so prepare
 session    optional     pam_ssh_add.so
 session    include      common-session
