+++ tools/debian/cockpit-bridge.install
@@ -1,7 +1,5 @@
 etc/cockpit/machines.d
 usr/bin/cockpit-bridge
-lib/*/security/pam_reauthorize.so
-usr/lib/cockpit/cockpit-polkit
 usr/lib/cockpit/cockpit-askpass
 usr/share/cockpit/base1/
 usr/share/man/man1/cockpit-bridge.1
