--- tools/debian/cockpit-bridge.lintian-overrides
@@ -1,1 +1,0 @@
-cockpit-bridge: setuid-binary usr/lib/cockpit/cockpit-polkit 4755 root/root
