+++ tools/debian/control
@@ -14,10 +14,8 @@ Build-Depends: debhelper (>= 9.20141010),
                zlib1g-dev,
                libkrb5-dev,
                libxslt1-dev,
-               libkeyutils-dev,
                libglib2.0-dev,
                libsystemd-dev,
-               libpolkit-agent-1-dev,
                libpcp3-dev,
                libjson-glib-dev,
                libpam0g-dev,
