+++ tools/semaphore-run
@@ -4,7 +4,6 @@ set -ex
 
 ./autogen.sh --prefix=/usr --enable-strict --with-systemdunitdir=/tmp
 make V=1 all
-sudo make enable-root-tests
 make -j8 distcheck
 make -j8 check-memory
 tools/static-code-tests
